+++ midap/apps/init_GUI.py
@@ -17,13 +17,18 @@ from midap.utils import get_logger, get_inheritors
 from midap.imcut import *
 from midap.imcut import base_cutout
 
-imcut_subclasses = [subclass.__name__ for subclass in get_inheritors(base_cutout.CutoutImage)]
+imcut_subclasses = [subclass for subclass in get_inheritors(base_cutout.CutoutImage)]
+family_imcut_cls = [s.__name__ for s in imcut_subclasses if "Family_Machine" in s.supported_setups]
+mother_imcut_cls = [s.__name__ for s in imcut_subclasses if "Mother_Machine" in s.supported_setups]
+
 
 # get all subclasses from the segmentations
 from midap.segmentation import *
 from midap.segmentation import base_segmentator
 
-segmentation_subclasses = [subclass.__name__ for subclass in get_inheritors(base_segmentator.SegmentationPredictor)]
+segmentation_subclasses = [subclass for subclass in get_inheritors(base_segmentator.SegmentationPredictor)]
+family_seg_cls = [s.__name__ for s in segmentation_subclasses if "Family_Machine" in s.supported_setups]
+mother_seg_cls = [s.__name__ for s in segmentation_subclasses if "Mother_Machine" in s.supported_setups]
 
 # get all subclasses from the tracking
 from midap.tracking import *
@@ -62,7 +67,7 @@ def main(config_file="settings.ini", loglevel=7):
     # First part of the GUI
     common_params = [[sg.Text("Select the input data type: ",
                               key="track_method_text", font="bold")],
-                     [sg.DropDown(key="DataType", values=["Family_Machine"], default_value="Family_Machine")],
+                     [sg.DropDown(key="DataType", values=["Family_Machine", "Mother_Machine"], default_value="Family_Machine")],
                      [sg.Text("Choose the target folder: ", key="title_folder_name", font="bold")],
                      [sg.Input(key="folder_name"), sg.FolderBrowse()],
                      [sg.Text("Filetype (e.g. tif, tiff, ome.tif)", key="title_file_type", font="bold")],
@@ -162,17 +167,28 @@ def main(config_file="settings.ini", loglevel=7):
                             [sg.Text("Thresholding: \n"
                                      "Enter a value between 0 (black) and 1 (white) to cap the brightest parts of the images.",
                                      font="bold")],
-                            [sg.Input(default_text=defaults["ImgThreshold"], size=30, key="thresholding_val")],
-                            
-                            # Segmentation
-                            [sg.Text("Segmentation options:",
-                                     font="bold")],
-                            [sg.Checkbox("Remove border cells", key="remove_border",
-                                         default=defaults.getboolean("RemoveBorder"), size=30)],
-                            ]
+                            [sg.Input(default_text=defaults["ImgThreshold"], size=30, key="thresholding_val")],]
+
+        if general["DataType"] == "Family_Machine":
+            # Segmentation
+            advanced_options += [[sg.Text("Segmentation options:", font="bold")],
+                                 [sg.Checkbox("Remove border cells", key="remove_border",
+                                         default=defaults.getboolean("RemoveBorder"), size=30)],]
+        if general["DataType"] == "Mother_Machine":
+            # mark cells on top or bottom of cells
+            advanced_options += [[sg.Text("During the tracking mark cell that are at the top/bottom of the chamber:", font="bold")],
+                                 [sg.DropDown(key="cell_marker", values=["top", "bottom", "both", "none"], default_value="none")]]
 
 
-        # Family Machine specific layout
+        # get the vars for the specific layout
+        if general["DataType"] == "Family_Machine":
+            imcut_subclasses = family_imcut_cls
+            segmentation_subclasses = family_seg_cls
+        if general["DataType"] == "Mother_Machine":
+            imcut_subclasses = mother_imcut_cls
+            segmentation_subclasses = mother_seg_cls
+
+        # Specific layout
         layout_family_machine = [[sg.Frame("Conditional Run", [[
             sg.Column(workflow, background_color="white"),
             sg.Column(frames)
@@ -200,7 +216,7 @@ def main(config_file="settings.ini", loglevel=7):
                                               default_value=defaults["TrackingClass"])],
                                  [sg.Text("Preprocessing", font="bold")],
                                  [sg.Checkbox("Deconvolution of images", key="deconv", font="bold",
-                                              default=(defaults["Deconvolution"] == "deconv_family_machine"))],
+                                              default=not (defaults["Deconvolution"] == "no_deconv"))],
                                  [sg.Text("")],
                                  [sg.Text(SYMBOL_RIGHT, enable_events=True, key='-OPEN_ADV-'),
                                   sg.Text("Advanced Options")],
@@ -256,7 +272,10 @@ def main(config_file="settings.ini", loglevel=7):
 
         # deconv
         if values["deconv"]:
-            section["Deconvolution"] = "deconv_family_machine"
+            if general["DataType"] == "Mother_Machine":
+                section["Deconvolution"] = "deconv_well"
+            elif general["DataType"] == "Family_Machine":
+                section["Deconvolution"] = "deconv_family_machine"
         else:
             section["Deconvolution"] = "no_deconv"
 
@@ -277,7 +296,11 @@ def main(config_file="settings.ini", loglevel=7):
         section["KeepSegImagesLabel"] = values["keep_seg_label"]
         section["KeepSegImagesBin"] = values["keep_seg_bin"]
         section["KeepSegImagesTrack"] = values["keep_seg_track"]
-        section["RemoveBorder"] = values["remove_border"]
+
+        if general["DataType"] == "Family_Machine":
+            section["RemoveBorder"] = values["remove_border"]
+        if general["DataType"] == "Mother_Machine":
+            section["CellMarker"] = values["cell_marker"]
 
         # Thresholding
         threshold = float(values["thresholding_val"])