+++ midap/segmentation/base_segmentator.py
@@ -96,9 +96,11 @@ class SegmentationPredictor(ABC):
             seg = label(seg, connectivity=self.connectivity)
 
             # save individual image
+            os.makedirs(path_seg, exist_ok=True)
             label_fname = re.sub("(_cut.tif|_cut.png|.tif)", "_seg.tif", p)
             io.imsave(os.path.join(path_seg, label_fname), seg.astype(np.uint16), check_contrast=False)
             seg_fname = re.sub("(_cut.tif|_cut.png|.tif)", "_seg_bin.png", p)
+            os.makedirs(path_seg_bin, exist_ok=True)
             io.imsave(os.path.join(path_seg_bin, seg_fname), 255*(seg > 0).astype(np.uint8), check_contrast=False)
 
     def postprocess_seg(self, seg: np.ndarray):