+++ midap/segmentation/unet_segmentator.py
@@ -19,6 +19,8 @@ class UNetSegmentation(SegmentationPredictor):
     A class that performs the image segmentation of the cells using a UNet
     """
 
+    supported_setups = ["Family_Machine"]
+
     def __init__(self, *args, **kwargs):
         """
         Initializes the UNetSegmentation using the base class init